Refactor build_info() python code per code review #5423
Many thanks to @zundertj for insightful comments/suggestions.
